make changes for RDFI vs ODFI
